REFACTOR: 로그 삭제 및 주석 생성

+++ src/main/java/com/meta/overwash/domain/LaundryPriceDTO.java
@@ -10,7 +10,7 @@ import lombok.ToString;
 @NoArgsConstructor
 @ToString
 public class LaundryPriceDTO {
-	private Integer priceId;
-	private Integer price;
+	private Long priceId;
+	private Long price;
 
 }
